TODO: Test against old HMM search code to make sure results are the same
