+++ requirements.txt
@@ -6,5 +6,5 @@ matplotlib
 mrcfile
 pandas
 fair-esm
-pyhmmer >= 0.7.1
+pyhmmer >= 0.10.1
 loguru
